+++ GV.DVDCentral.DB/Tables/tblOrderItem.sql
@@ -4,5 +4,5 @@
 	[OrderId] INT NOT NULL,
 	[MovieId] INT NOT NULL,
 	[Quantity] INT NOT NULL,
-	[Cost] DECIMAL(18,2) NOT NULL
+	[Cost] FLOAT NOT NULL
 )
